Added container class and restructured directory
