+++ src/bin/facilitator-start.ts
@@ -1,8 +1,7 @@
 import * as commander from 'commander';
 import Logger from '../Logger';
 import Facilitator from '../Facilitator';
-import { Config } from '../Config';
-import FacilitatorStart from '../OptionParser/FacilitatorStart';
+import Container from '../Container';
 
 const facilitatorCmd = commander
   .arguments('[origin_chain] [aux_chain_id]');
@@ -23,17 +22,13 @@ facilitatorCmd
   .option('-mc, --mosaic-config <mosaicConfig>', 'path to mosaic configuration')
   .option('-fc, --facilitator-config <facilitatorConfig>', 'path to facilitator configuration')
   .action(async (origin_chain, aux_chain_id, options) => {
-    let configObj: Config;
-
     try {
-      const facilitatorStart: FacilitatorStart = new FacilitatorStart(
+      facilitator = new Container(
         origin_chain,
         aux_chain_id,
         options.mosaicConfig,
         options.facilitatorConfig,
-      );
-      configObj = facilitatorStart.getConfig();
-      facilitator = new Facilitator(configObj);
+      ).construct();
 
       Logger.info('starting facilitator');
       await facilitator.start();
