+++ src/services/Services.ts
@@ -1,0 +1,3 @@
+export default class Services {
+
+}
