+++ src/subscriptions/Subscriber.ts
@@ -1,8 +1,8 @@
 import { Subscription } from 'apollo-client/util/Observable';
 import GraphClient from './GraphClient';
-import TransactionHandler from './TransactionHandler';
+import TransactionHandler from '../TransactionHandler';
 import TransactionFetcher from './TransactionFetcher';
-import ContractEntityRepository from './repositories/ContractEntityRepository';
+import ContractEntityRepository from '../repositories/ContractEntityRepository';
 
 /**
  * Subscriber class subscribes and unsubscribes subscription queries of a subgraph.
