+++ test/subscriptions/Subscriber/unsusbcribe.test.ts
@@ -1,31 +1,30 @@
 import * as sinon from 'sinon';
-import assert from '../test_utils/assert';
+import assert from '../../test_utils/assert';
 
-import Subscriber from '../../src/Subscriber';
-import GraphClient from '../../src/GraphClient';
-import SpyAssert from '../test_utils/SpyAssert';
-import TransactionHandler from '../../src/TransactionHandler';
-import TransactionFetcher from '../../src/TransactionFetcher';
-import ContractEntityRepository from '../../src/repositories/ContractEntityRepository';
+import Subscriber from '../../../src/subscriptions/Subscriber';
+import GraphClient from '../../../src/subscriptions/GraphClient';
+import TransactionHandler from '../../../src/TransactionHandler';
+import TransactionFetcher from '../../../src/subscriptions/TransactionFetcher';
+import ContractEntityRepository from '../../../src/repositories/ContractEntityRepository';
 
-describe('Subscriber.subscribe()', () => {
+describe('Subscriber.unsubscribe()', () => {
   let mockApolloClient: any;
   let graphClient: GraphClient;
   let subscriptionQueries: Record<string, string>;
   let subscriber: Subscriber;
+  let mockUnsubscribe: any;
 
   beforeEach(() => {
     mockApolloClient = sinon.stub;
     graphClient = new GraphClient(mockApolloClient);
     subscriptionQueries = { stakeRequesteds: 'subscription{stakeRequesteds{id}}' };
-  });
-
-  it('should work with correct parameters', async () => {
-    const mockQuerySubscriber = sinon.spy as any;
-    const spyGraphClientSubscribe = sinon.replace(
+    mockUnsubscribe = {
+      unsubscribe: sinon.spy,
+    };
+    sinon.replace(
       graphClient,
       'subscribe',
-      sinon.fake.resolves(mockQuerySubscriber),
+      sinon.fake.resolves(mockUnsubscribe),
     );
     const handler = sinon.mock(TransactionHandler);
     const fetcher = sinon.mock(TransactionFetcher);
@@ -37,6 +36,9 @@ describe('Subscriber.subscribe()', () => {
       fetcher as any,
       contractEntityRepository as any,
     );
+  });
+
+  it('should work with correct parameters', async () => {
     await subscriber.subscribe();
 
     assert.strictEqual(
@@ -45,16 +47,18 @@ describe('Subscriber.subscribe()', () => {
       'Subscription failed.',
     );
 
-    assert.strictEqual(
+    const mockQuerySubscription = sinon.spy;
+    sinon.replace(
       subscriber.querySubscriptions.stakeRequesteds,
-      mockQuerySubscriber,
-      'Invalid query subscription object.',
+      'unsubscribe',
+      sinon.fake.resolves(mockQuerySubscription),
     );
 
-    SpyAssert.assert(
-      spyGraphClientSubscribe,
-      1,
-      [[subscriptionQueries.stakeRequesteds, handler, fetcher, contractEntityRepository]],
+    await subscriber.unsubscribe();
+    assert.strictEqual(
+      Object.keys(subscriber.querySubscriptions).length,
+      0,
+      'UnSubscription failed.',
     );
 
     sinon.restore();
